feat: relocate CalendarSidebar to components/dashboard directory for better organization
